requirements.txt
aiohttp==3.7.4.post0
altgraph==0.17
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.1
discord.py==1.7.3
future==0.18.2
idna==3.2
multidict==5.1.0
pefile==2021.5.24
pyinstaller==4.4
pyinstaller-hooks-contrib==2021.2
pywin32-ctypes==0.2.0
requests==2.26.0
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3
